@@ -1,8 +1,8 @@
-import React, { Component } from 'react';
+import React from 'react';
 import {
   BrowserRouter as Router,
+  Routes,
   Route,
-  Switch,
 } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import AboutPage from './pages/AboutPage';
@@ -12,25 +12,23 @@ import NotFoundPage from './pages/NotFoundPage';
 import NavBar from './NavBar';
 import './App.css';
 
-class App extends Component {
-  render() {
-    return (
-      <Router>
-        <div className="App">
-          <NavBar />
-          <div id="page-body">
-            <Switch>
-              <Route path="/" component={HomePage} exact />
-              <Route path="/about" component={AboutPage} />
-              <Route path="/articles-list" component={ArticlesListPage} />
-              <Route path="/article/:name" component={ArticlePage} />
-              <Route component={NotFoundPage} />
-            </Switch>
-          </div>
+function App() {
+  return (
+    <Router>
+      <div className="App">
+        <NavBar />
+        <div id="page-body">
+          <Routes>
+            <Route path="/" element={<HomePage />} />
+            <Route path="/about" element={<AboutPage />} />
+            <Route path="/articles-list" element={<ArticlesListPage />} />
+            <Route path="/article/:name" element={<ArticlePage />} />
+            <Route path="*" element={<NotFoundPage />} />
+          </Routes>
         </div>
-      </Router>
-    );
-  }
+      </div>
+    </Router>
+  );
 }
 
 export default App;
